requirements.txt
# BILK Medical Robotics System - Python Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies for host bridge and simulation
numpy>=1.21.0              # Numerical computing for smoothing and filtering
matplotlib>=3.5.0          # Plotting and visualization for diagnostics
pyserial>=3.5              # Serial communication with Arduino/ESP32

# Optional dependencies for enhanced functionality
scipy>=1.7.0               # Scientific computing (for advanced filtering)
pandas>=1.3.0              # Data analysis (for performance metrics)
seaborn>=0.11.0            # Statistical visualization (for diagnostic plots)

# Development and testing dependencies (optional)
pytest>=6.0.0              # Testing framework
pytest-cov>=2.12.0         # Coverage reporting
black>=21.0.0              # Code formatting
flake8>=3.9.0              # Linting

# System monitoring (optional)
psutil>=5.8.0              # System resource monitoring
